+++ client/Packages/com.beamable/CHANGELOG.md
@@ -12,7 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Editor tooling for `SerializableDictionaryStringToSomething<T>` has context menu
-
+- `ClearCaches` function on `StatsApi` to force invalidate stats cache.
 
 ### [1.3.2]